requirements.txt
blinker
certifi
charset-normalizer
click
dnspython
flask
Flask-Cors
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
pymongo
python-dateutil
python-dotenv
pytz
requests
six
tzdata
urllib3
werkzeug
zipp
gunicorn